Single window shell experiment
A single window shell is easier because we don't have to worry about
multiple windows, however this doesn't work with Weston as opposed to
qtwayland.  With the QtCompositor API we can set the shell window as
parent of all surfaces and play with z-index while on Weston we have
the shell window on the background layer and surfaces on top of it.